put js for side panels in seperate file
